tela de Ecopoints cadasto e Listagem

+++ api/models/EcopontoModel.cs
@@ -29,9 +29,8 @@ namespace api.Models
 
         public int CEP { get; set; }
 
-        public int Latitude { get; set; }
-
-        public int Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
 
 
         public string Username { get; set; } = string.Empty;
